added max_tokens function to llm.py, made separate file for skip dir+file
