+++ config/config.yml
@@ -1,3 +1,5 @@
+BASE_MAX_TOKENS_CODE: 90
+BASE_MAX_TOKENS_SUM: 120
 MAX_TOKENS_CODE: 90
 MAX_TOKENS_SUM: 160
 MAX_TOKENS_CHAIN: 512
