+++ src/llm.py
@@ -6,6 +6,7 @@ from langchain_openai import OpenAI, ChatOpenAI
 from src.llm_chains import chain_summarize_summaries
 from src.prompt_templates import code_summary_prompt, summaries_summary_prompt
 from src.chat_prompt_templates import chat_code_summary_prompt, chat_sum_summary_prompt
+from tokens_weighting import max_num_tokens
 
 #load API key from .env
 load_dotenv()
@@ -17,11 +18,12 @@ with open('config/config.yml', 'r', encoding='utf8') as ymlfile:
 
 def llm_generate_summary(file_name: str, code: str) -> str:
     """Generate a summary of the code, using an LLM."""
+    max_tokens = max_num_tokens(file_name, cfg.BASE_MAX_TOKENS_CODE)
     if cfg.MODEL_TYPE == "completion":
-        llm = create_llm(cfg.MAX_TOKENS_CODE)
+        llm = create_llm(max_tokens=max_tokens)
         prompt = code_summary_prompt(file_name, code)
     else:
-        llm = create_chat_llm(cfg.MAX_TOKENS_CODE)
+        llm = create_chat_llm(max_tokens=max_tokens)
         prompt = chat_code_summary_prompt(file_name, code)
     if is_prompt_too_big(llm, prompt):
         output = chain_summarize_summaries(code)  # FIXME: What is the prompt used here?
@@ -32,11 +34,12 @@ def llm_generate_summary(file_name: str, code: str) -> str:
 
 def llm_summarize_summary(component: str, summaries: list[str]) -> str:
     """Generate a summary of summaries, using an LLM."""
+    max_tokens = max_num_tokens(component, cfg.BASE_MAX_TOKENS_SUM)
     if cfg.MODEL_TYPE == "completion":
-        llm = create_llm(cfg.MAX_TOKENS_SUM)
+        llm = create_llm(max_tokens=max_tokens)
         prompt = summaries_summary_prompt(component, summaries)
     else:
-        llm = create_chat_llm(cfg.MAX_TOKENS_SUM)
+        llm = create_chat_llm(max_tokens=max_tokens)
         prompt = chat_sum_summary_prompt(component, summaries)
     if is_prompt_too_big(llm, prompt):
         output = chain_summarize_summaries(summaries)  # FIXME: What is the prompt used here?
